Add ESP32-C6-DevkitC-1 symbol and footprint (#103)
* Add ESP32-C6-DevKitC-1 symbol

Based on the ESP32-S2-DevKitC-1

* Add ESP32-C6-DevKitC-1 footprint

This is derived from the ESP32-S2-DevKitC-1 footprint to keep style and
addapted as best effort in relation to the dimensional drawings and a
simplified version of the dxf provided with the espressif documentation.

I am not so sure about the tilde suffixes used in the S2 footprint
design, I think they are meant to mean PWM capable outputs. Possibly and
Arduinoism. I added them on essentially all pins and this what the
espressif documentation stats as PWM capable pins,... but yet they have
it on the 5V pin.

* Make ESP32-C6-DevKitC-1 symbol NC pins unconnected pin type

* ESP32-C6-DevKitC-1 symbol add alternate pin definitions

- Use GPIOx/ notation on all alternate pins
- Keep strapping pin functions as defaults, plus for USB and UART0
- Fixed a typo for GPIO22 on pin 27